+++ packages/terra-layout/examples/test-examples/LayoutNoHeader.jsx
@@ -5,6 +5,7 @@ import TestContent from './test-content/TestContent';
 
 const LayoutNoHeader = () => (
   <Layout
+    style={{ height: '768px' }}
     key="layout-no-header"
     menu={<TestMenu />}
     menuText="Menu Text"