+++ packages/terra-application-links/src/tabs/_TabUtils.js
@@ -1,0 +1,9 @@
+const KEYCODES = {
+  ENTER: 13,
+};
+
+const TabUtils = {
+  KEYCODES,
+};
+
+export default TabUtils;