+++ packages/terra-modal-manager/src/index.jsx
@@ -1,7 +1,5 @@
-import { connect } from 'react-redux';
-
-import ModalManager from './ModalManager';
-
+import ModalManager, { disclosureType } from './ModalManager';
+import withModalManager from './withModalManager';
 import modalManagerReducers from './reducers';
 import { open, close, push, pop, maximize, minimize, gainFocus, loseFocus } from './actions';
 
@@ -30,9 +28,12 @@ const mapDispatchToProps = dispatch => ({
 
 export { mapDispatchToProps };
 
-export default connect(mapStateToProps, mapDispatchToProps)(ModalManager);
-
 const reducers = {
   modalManager: modalManagerReducers,
 };
 export { reducers };
+
+export default ModalManager;
+export { disclosureType };
+
+export { withModalManager };